* Refactor flowRunner so that test artifact doesn't bring its transitive dependencies
